move save_checkpoint
(cherry picked from commit 222102c81471874d941303b17ca8859d11410552)
